requirements.txt
alibabacloud-alidns20150109
PyYAML
requests
